rename .es > .js
