+++ index.js
@@ -1,2 +1,2 @@
 'use strict';
-module.exports = require('./lib/deepcrawl.es');
+module.exports = require('./lib/deepcrawl');
