+++ lib/deepcrawl.js
@@ -1,6 +1,6 @@
 'use strict';
 const BPromise = require('bluebird'),
-  schema = require('./schema.es'),
+  schema = require('./schema'),
   needle = BPromise.promisifyAll(require('needle')),
   humps = require('humps'),
   _ = require('lodash');
