+++ lib/schema.js
